Use base entity
